requirements.txt
PyYAML==6.0.1
click==8.1.3
requests==2.28.1
tzdata==2022.2
